+++ src/pages/MainPage.tsx
@@ -1,14 +1,14 @@
-import CardList from '../components/mainpage/CardList';
+import PopularCardList from '../components/mainpage/PopularCardList';
 
 const MainPage = () => {
   return (
     <div className="flex flex-col items-center w-full p-12">
-      <CardList category="painting" />
-      <CardList category="music" />
-      <CardList category="exercise" />
-      <CardList category="cooking" />
-      <CardList category="game" />
-      <CardList category="movie" />
+      <PopularCardList category="painting" />
+      <PopularCardList category="music" />
+      <PopularCardList category="exercise" />
+      <PopularCardList category="cooking" />
+      <PopularCardList category="game" />
+      <PopularCardList category="movie" />
     </div>
   );
 };
